feat: chat, config 분리
